PedagogicalFlowExtractor/requirements.txt
# ─── Core Dependencies (MVP) ───
openai-whisper>=20230314
torch>=2.0.0
ffmpeg-python>=0.2.0
pydub>=0.25.1

# NLP
rake-nltk>=1.0.6
nltk>=3.8.0

# Graph
networkx>=3.0
pyvis>=0.3.1

# Visualization
plotly>=5.14.0

# Web App
streamlit>=1.28.0

# Config & Environment
pyyaml>=6.0
python-dotenv>=1.0.0

# ─── Enhanced (Phase 2) ───
groq>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
yt-dlp>=2024.1.0
